Enable -Werror and fix all errors

+++ veranda/Packages/qt_frontend/builtin_qt_frontend_plugins/pkg-veranda_qt_frontend_image_file_handler/src/imageparser.cpp
@@ -64,7 +64,7 @@ QVector<Shape> ImageParser::parseImage(const QVector<QVector<QRgb> >& pixMap, co
     {
         for(int j=0; j<pixMap[i].size(); j++)
         {
-            blackWhite[i][j] = qGray(pixMap[i][j]) > colorThreshold;
+            blackWhite[i][j] = static_cast<uint64_t>(qGray(pixMap[i][j])) > colorThreshold;
         }
     }
